requirements.txt
albumentations
efficientnet_pytorch
vit-pytorch
seaborn
notion
dropbox
requests
fastprogress
easydict
PyYAML
kaggle
optuna
joblib
easydict
GitPython
rfconv
neptune-client
psutil